backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-dotenv
python-jose[cryptography]
email-validator   ✅ add this